+++ src/testSetup.ts
@@ -15,7 +15,6 @@ import { tmpdir as osTmpdir } from 'os';
 import { basename, join as pathJoin, dirname } from 'path';
 import { SinonSandbox, SinonStatic, SinonStub } from 'sinon';
 
-import { stubMethod } from '@salesforce/ts-sinon';
 import {
   AnyFunction,
   AnyJson,
@@ -290,9 +289,12 @@ export class TestContext {
     );
     const orgMap = new Map(entries);
 
-    stubMethod(this.SANDBOX, OrgAccessor.prototype, 'getAllFiles').resolves([...orgMap.keys()].map((o) => `${o}.json`));
+    // @ts-expect-error because private method
+    this.SANDBOX.stub(OrgAccessor.prototype, 'getAllFiles').resolves([...orgMap.keys()].map((o) => `${o}.json`));
 
-    stubMethod(this.SANDBOX, OrgAccessor.prototype, 'hasFile').callsFake((username: string) => orgMap.has(username));
+    this.SANDBOX.stub(OrgAccessor.prototype, 'hasFile').callsFake((username: string) =>
+      Promise.resolve(orgMap.has(username))
+    );
 
     const retrieveContents = async function (this: { path: string }): Promise<AuthFields> {
       const username = basename(this.path.replace('.json', ''));
@@ -337,9 +339,11 @@ export class TestContext {
       })
     );
 
-    stubMethod(this.SANDBOX, User.prototype, 'retrieve').callsFake(
-      (username): Promise<UserFields | undefined> => Promise.resolve(mockUsers.find((org) => org.username === username))
-    );
+    this.SANDBOX.stub(User.prototype, 'retrieve').callsFake((username): Promise<UserFields> => {
+      const user = mockUsers.find((org) => org.username === username);
+      if (!user) throw new SfError('User not found', 'UserNotFoundError');
+      return Promise.resolve(user);
+    });
 
     const retrieveContents = async function (this: { path: string }): Promise<{ usernames?: string[] }> {
       const orgId = basename(this.path.replace('.json', ''));
@@ -358,7 +362,8 @@ export class TestContext {
     )) as Array<[string, SandboxFields]>;
     const sandboxMap = new Map(entries);
 
-    stubMethod(this.SANDBOX, SandboxAccessor.prototype, 'getAllFiles').resolves(
+    // @ts-expect-error because private method
+    this.SANDBOX.stub(SandboxAccessor.prototype, 'getAllFiles').resolves(
       [...sandboxMap.keys()].map((o) => `${o}.sandbox.json`)
     );
 
@@ -501,8 +506,8 @@ export const stubContext = (testContext: TestContext): Record<string, SinonStub>
   const stubs: Record<string, SinonStub> = {};
 
   // Most core files create a child logger so stub this to return our test logger.
-  stubMethod(testContext.SANDBOX, Logger, 'child').resolves(testContext.TEST_LOGGER);
-  stubMethod(testContext.SANDBOX, Logger, 'childFromRoot').returns(testContext.TEST_LOGGER);
+  testContext.SANDBOX.stub(Logger, 'child').resolves(testContext.TEST_LOGGER);
+  testContext.SANDBOX.stub(Logger, 'childFromRoot').returns(testContext.TEST_LOGGER);
   testContext.inProject(true);
   testContext.SANDBOXES.CONFIG.stub(ConfigFile, 'resolveRootFolder').callsFake((isGlobal: boolean) =>
     testContext.rootPathRetriever(isGlobal, testContext.id)
@@ -511,7 +516,8 @@ export const stubContext = (testContext: TestContext): Record<string, SinonStub>
     testContext.rootPathRetrieverSync(isGlobal, testContext.id)
   );
 
-  stubMethod(testContext.SANDBOXES.PROJECT, SfProjectJson.prototype, 'doesPackageExist').callsFake(() => true);
+  // @ts-expect-error using private method
+  testContext.SANDBOXES.PROJECT.stub(SfProjectJson.prototype, 'doesPackageExist').callsFake(() => true);
 
   const initStubForRead = (configFile: ConfigFile<ConfigFile.Options>): ConfigStub => {
     testContext.configStubs[configFile.constructor.name] ??= {};
@@ -568,13 +574,13 @@ export const stubContext = (testContext: TestContext): Record<string, SinonStub>
     writeSync.call(this);
   };
 
-  stubs.configWriteSync = stubMethod(testContext.SANDBOXES.CONFIG, ConfigFile.prototype, 'writeSync').callsFake(
-    writeSync
-  );
+  stubs.configWriteSync = testContext.SANDBOXES.CONFIG.stub(ConfigFile.prototype, 'writeSync').callsFake(writeSync);
 
-  stubs.configWrite = stubMethod(testContext.SANDBOXES.CONFIG, ConfigFile.prototype, 'write').callsFake(write);
+  stubs.configWrite = testContext.SANDBOXES.CONFIG.stub(ConfigFile.prototype, 'write').callsFake(write);
 
-  stubMethod(testContext.SANDBOXES.CRYPTO, Crypto.prototype, 'getKeyChain').callsFake(() =>
+  // @ts-expect-error: getKeyChain is private
+  testContext.SANDBOXES.CRYPTO.stub(Crypto.prototype, 'getKeyChain').callsFake(() =>
+    // @ts-expect-error: not the full type
     Promise.resolve({
       setPassword: () => Promise.resolve(),
       getPassword: (data: Record<string, unknown>, cb: AnyFunction) =>
@@ -582,9 +588,10 @@ export const stubContext = (testContext: TestContext): Record<string, SinonStub>
     })
   );
 
-  stubMethod(testContext.SANDBOXES.CONNECTION, Connection.prototype, 'isResolvable').resolves(true);
+  testContext.SANDBOXES.CONNECTION.stub(Connection.prototype, 'isResolvable').resolves(true);
 
-  stubMethod(testContext.SANDBOXES.CONNECTION, Connection.prototype, 'request').callsFake(function (
+  // @ts-expect-error: just enough of an httpResponse for testing
+  testContext.SANDBOXES.CONNECTION.stub(Connection.prototype, 'request').callsFake(function (
     this: Connection,
     request: string,
     options?: Dictionary
@@ -595,23 +602,25 @@ export const stubContext = (testContext: TestContext): Record<string, SinonStub>
     return testContext.fakeConnectionRequest.call(this, request, options as AnyJson);
   });
 
-  stubMethod(testContext.SANDBOX, aliasAccessorEntireFile, 'getFileLocation').returns(getAliasFileLocation());
+  testContext.SANDBOX.stub(aliasAccessorEntireFile, 'getFileLocation').returns(getAliasFileLocation());
 
-  stubs.configExists = stubMethod(testContext.SANDBOXES.ORGS, OrgAccessor.prototype, 'exists').callsFake(
-    async function (this: OrgAccessor, username: string): Promise<boolean | undefined> {
-      // @ts-expect-error because private member
-      if ([...this.contents.keys()].includes(username)) return Promise.resolve(true);
-      else return Promise.resolve(false);
-    }
-  );
+  stubs.configExists = testContext.SANDBOXES.ORGS.stub(OrgAccessor.prototype, 'exists').callsFake(async function (
+    this: OrgAccessor,
+    username: string
+  ): Promise<boolean> {
+    // @ts-expect-error because private member
+    if ([...this.contents.keys()].includes(username)) return Promise.resolve(true);
+    else return Promise.resolve(false);
+  });
 
-  stubs.configRemove = stubMethod(testContext.SANDBOXES.ORGS, OrgAccessor.prototype, 'remove').callsFake(
-    async function (this: OrgAccessor, username: string): Promise<boolean | undefined> {
-      // @ts-expect-error because private member
-      if ([...this.contents.keys()].includes(username)) return Promise.resolve(true);
-      else return Promise.resolve(false);
-    }
-  );
+  stubs.configRemove = testContext.SANDBOXES.ORGS.stub(OrgAccessor.prototype, 'remove').callsFake(async function (
+    this: OrgAccessor,
+    username: string
+  ): Promise<void> {
+    // @ts-expect-error because private member
+    if ([...this.contents.keys()].includes(username)) return Promise.resolve();
+    else return Promise.resolve();
+  });
 
   // Always start with the default and tests beforeEach or it methods can override it.
   testContext.fakeConnectionRequest = defaultFakeConnectionRequest;